tests/test_radon.py
import unittest

import numpy as np
from skimage.transform import radon as sk_radon, iradon as sk_iradon
import torch

from pytorch_radon import Radon, IRadon


class TestRadon(unittest.TestCase):
    def test_radon_iradon_circle(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = True
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle)
        ir = IRadon(img.shape[2], theta, circle)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img, reco).item(),
            0,
            places=4,
        )

    def test_radon_iradon_circle_cut_output(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = True
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle)
        ir = IRadon(img.shape[2], theta, circle, out_size=128)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img[:, :, 64:192, 64:192], reco).item(),
            0,
            places=3,
        )

    def test_radon_iradon_circle_lazy(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = True
        theta = torch.arange(180)
        r = Radon(theta=theta, circle=circle)
        ir = IRadon(theta=theta, circle=circle)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img, reco).item(),
            0,
            places=3,
        )

    def test_radon_iradon_circle_lazy_cut_output(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = True
        theta = torch.arange(180)
        r = Radon(theta=theta, circle=circle)
        ir = IRadon(theta=theta, circle=circle, out_size=128)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img[:, :, 64:192, 64:192], reco).item(),
            0,
            places=3,
        )

    def test_radon_iradon_not_circle(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = False
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle)
        ir = IRadon(img.shape[2], theta, circle)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img, reco).item(),
            0,
            places=4,
        )

    def test_radon_iradon_not_circle_cut_output(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = False
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle)
        ir = IRadon(img.shape[2], theta, circle, out_size=128)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img[:, :, 64:192, 64:192], reco).item(),
            0,
            places=3,
        )

    def test_radon_iradon_not_circle_lazy(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = False
        theta = torch.arange(180)
        r = Radon(theta=theta, circle=circle)
        ir = IRadon(theta=theta, circle=circle)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img, reco).item(),
            0,
            places=3,
        )

    def test_radon_iradon_not_circle_lazy_cut_output(self):
        img = torch.zeros(1, 1, 256, 256)
        img[:, :, 120:130, 120:130] = 1
        circle = False
        theta = torch.arange(180)
        r = Radon(theta=theta, circle=circle)
        ir = IRadon(theta=theta, circle=circle, out_size=128)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img[:, :, 64:192, 64:192], reco).item(),
            0,
            places=3,
        )

    def test_radon_iradon_circle_double(self):
        img = torch.zeros(1, 1, 256, 256, dtype=torch.double)
        img[:, :, 120:130, 120:130] = 1
        circle = True
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle, dtype=torch.double)
        ir = IRadon(img.shape[2], theta, circle, dtype=torch.double)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img, reco).item(),
            0,
            places=4,
        )
        self.assertTrue(reco.dtype == torch.double)

    def test_radon_iradon_not_circle_double(self):
        img = torch.zeros(1, 1, 256, 256, dtype=torch.double)
        img[:, :, 120:130, 120:130] = 1
        circle = False
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle, dtype=torch.double)
        ir = IRadon(img.shape[2], theta, circle, dtype=torch.double)
        sino = r(img)
        reco = ir(sino)
        self.assertAlmostEqual(
            torch.nn.MSELoss()(img, reco).item(),
            0,
            places=4,
        )
        self.assertTrue(reco.dtype == torch.double)

    def test_radon_iradon_not_circle_double_scikit(self):
        img = torch.zeros(1, 1, 256, 256, dtype=torch.double)
        img[:, :, 120:130, 120:130] = 1
        circle = False
        theta = torch.arange(180)
        r = Radon(img.shape[2], theta, circle, dtype=torch.double, scikit=True)
        ir = IRadon(
            img.shape[2],
            theta,
            circle,
            dtype=torch.double,
            use_filter=None,
            scikit=True,
        )
        sino = r(img)
        reco = ir(sino)

        sino_sk = sk_radon(
            img.numpy().squeeze(),
            theta=theta.numpy(),
            circle=circle,
        )
        reco_sk = sk_iradon(
            sino_sk,
            theta=theta.numpy(),
            circle=circle,
            filter_name=None,
        )

        self.assertAlmostEqual(
            np.mean((sino_sk - sino.numpy().squeeze())**2),
            0,
            places=27,
        )

        # allowed to fail until fixed
        self.assertNotAlmostEqual(
            np.mean((reco_sk - reco.numpy().squeeze())**2),
            0,
            places=28,
        )

    def test_radon_scikit_even(self):
        img = np.zeros((4, 4))
        img[1, 2] = 1.0
        img[2, 2] = 1.0
        timg = torch.from_numpy(img).float().cuda()[None, None, ...]
        theta = list(np.linspace(0, 180, 180, endpoint=False))
        tradon = Radon(theta=theta, scikit=True)
        sino = sk_radon(img, theta=theta, circle=True)
        tsino = tradon(timg)
        self.assertAlmostEqual(
            np.mean((sino - tsino.cpu().numpy().squeeze())**2),
            0,
            places=13,
        )

    def test_radon_scikit_odd(self):
        img = np.zeros((5, 5))
        img[1, 2] = 1.0
        img[2, 2] = 1.0
        timg = torch.from_numpy(img).float().cuda()[None, None, ...]
        theta = list(np.linspace(0, 180, 180, endpoint=False))
        tradon = Radon(theta=theta, scikit=True)
        sino = sk_radon(img, theta=theta, circle=True)
        tsino = tradon(timg)
        self.assertAlmostEqual(
            np.mean((sino - tsino.cpu().numpy().squeeze())**2),
            0,
            places=13,
        )

    def test_iradon_scikit_odd(self):
        img = np.zeros((5, 5))
        img[1, 2] = 1.0
        img[2, 2] = 1.0
        timg = torch.from_numpy(img).float().cuda()[None, None, ...]
        theta = list(np.linspace(0, 180, 180, endpoint=False))
        tradon = Radon(theta=theta, scikit=True)
        tsino = tradon(timg)

        tiradon = IRadon(theta=theta, use_filter=None, scikit=True)
        reco = sk_iradon(tsino.cpu().numpy()[0, 0], theta=theta, filter_name=None)
        treco = tiradon(tsino)

        self.assertAlmostEqual(
            np.mean((reco - treco.cpu().numpy().squeeze())**2),
            0,
            places=13,
        )

    def test_iradon_scikit_even(self):
        img = np.zeros((4, 4))
        img[1, 2] = 1
        img[2, 2] = 1
        timg = torch.from_numpy(img).float().cuda()[None, None, ...]
        theta = list(np.linspace(0, 180, 4, endpoint=False))
        tradon = Radon(theta=theta, scikit=True)
        tsino = tradon(timg)

        with self.assertWarns(Warning):
            tiradon = IRadon(theta=theta, use_filter=None, scikit=True)
            reco = sk_iradon(tsino.cpu().numpy()[0, 0], theta=theta, filter_name=None)
            treco = tiradon(tsino)

            # allowed to fail until fixed
            self.assertNotAlmostEqual(
                np.mean((reco - treco.cpu().numpy().squeeze())**2),
                0,
                places=13,
            )


if __name__ == '__main__':
    unittest.main()
